requirements.txt
numpy==2.3.0
pyyaml
biopython==1.78
# matplotlib==3.4.3
patchworklib==0.4.5
python-queen==1.1.0
seaborn==0.13.2
